+++ src/parser/atomizer/AtomizerFactory.cpp
@@ -30,7 +30,8 @@
 
 namespace Opal {
 
-std::vector<std::unique_ptr<AtomizerBase>> AtomizerFactory::createAtomizers(int& current, std::vector<Token>& tokens) {
+std::vector<std::unique_ptr<AtomizerBase>> AtomizerFactory::createAtomizers(size_t&             current,
+                                                                            std::vector<Token>& tokens) {
     std::vector<std::unique_ptr<AtomizerBase>> atomizers;
 
     atomizers.push_back(std::make_unique<VariableAtomizer>(current, tokens));
